move main conflict helpers to their own file
